tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.roundrect.1.radius.dompointinit.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.roundrect.1.radius.dompointinit</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.roundrect.1.radius.dompointinit</h1>
<p class="desc">Verify that when one radius is given to roundRect(), specified as a DOMPointInit, applies to all corners.</p>


<script>
var t = async_test("Verify that when one radius is given to roundRect(), specified as a DOMPointInit, applies to all corners.");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.roundRect(0, 0, 100, 50, [{x: 40, y: 20}]);
  ctx.fillStyle = '#0f0';
  ctx.fill();

  // top-left corner
  _assertPixel(canvas, 20,1, 255,0,0,255);
  _assertPixel(canvas, 41,1, 0,255,0,255);
  _assertPixel(canvas, 1,10, 255,0,0,255);
  _assertPixel(canvas, 1,21, 0,255,0,255);

  // top-right corner
  _assertPixel(canvas, 79,1, 255,0,0,255);
  _assertPixel(canvas, 58,1, 0,255,0,255);
  _assertPixel(canvas, 98,10, 255,0,0,255);
  _assertPixel(canvas, 98,21, 0,255,0,255);

  // bottom-right corner
  _assertPixel(canvas, 79,48, 255,0,0,255);
  _assertPixel(canvas, 58,48, 0,255,0,255);
  _assertPixel(canvas, 98,39, 255,0,0,255);
  _assertPixel(canvas, 98,28, 0,255,0,255);

  // bottom-left corner
  _assertPixel(canvas, 20,48, 255,0,0,255);
  _assertPixel(canvas, 41,48, 0,255,0,255);
  _assertPixel(canvas, 1,39, 255,0,0,255);
  _assertPixel(canvas, 1,28, 0,255,0,255);
  t.done();

});
</script>
